+++ ee/tabby-ui/components/chat/chat-scroll-anchor.tsx
