+++ ee/tabby-ui/components/chat/empty-screen.tsx
