+++ ee/tabby-ui/components/chat/prompt-form.tsx
